Moved some UTs
