+++ pkg/injector/sidecar/container_test.go
@@ -18,9 +18,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"github.com/dapr/dapr/pkg/injector/annotations"
 	authConsts "github.com/dapr/dapr/pkg/runtime/security/consts"
@@ -31,103 +29,6 @@ const (
 	defaultAPITokenSecret = "secret"
 	defaultAppTokenSecret = "appsecret"
 )
-
-func TestGetResourceRequirements(t *testing.T) {
-	t.Run("no resource requirements", func(t *testing.T) {
-		r, err := getResourceRequirements(nil)
-		require.NoError(t, err)
-		assert.Nil(t, r)
-	})
-
-	t.Run("valid resource limits", func(t *testing.T) {
-		a := map[string]string{annotations.KeyCPULimit: "100m", annotations.KeyMemoryLimit: "1Gi"}
-		r, err := getResourceRequirements(a)
-		require.NoError(t, err)
-		assert.Equal(t, "100m", r.Limits.Cpu().String())
-		assert.Equal(t, "1Gi", r.Limits.Memory().String())
-	})
-
-	t.Run("invalid cpu limit", func(t *testing.T) {
-		a := map[string]string{annotations.KeyCPULimit: "cpu", annotations.KeyMemoryLimit: "1Gi"}
-		r, err := getResourceRequirements(a)
-		assert.NotNil(t, err)
-		assert.Nil(t, r)
-	})
-
-	t.Run("invalid memory limit", func(t *testing.T) {
-		a := map[string]string{annotations.KeyCPULimit: "100m", annotations.KeyMemoryLimit: "memory"}
-		r, err := getResourceRequirements(a)
-		assert.NotNil(t, err)
-		assert.Nil(t, r)
-	})
-
-	t.Run("valid resource requests", func(t *testing.T) {
-		a := map[string]string{annotations.KeyCPURequest: "100m", annotations.KeyMemoryRequest: "1Gi"}
-		r, err := getResourceRequirements(a)
-		require.NoError(t, err)
-		assert.Equal(t, "100m", r.Requests.Cpu().String())
-		assert.Equal(t, "1Gi", r.Requests.Memory().String())
-	})
-
-	t.Run("invalid cpu request", func(t *testing.T) {
-		a := map[string]string{annotations.KeyCPURequest: "cpu", annotations.KeyMemoryRequest: "1Gi"}
-		r, err := getResourceRequirements(a)
-		assert.NotNil(t, err)
-		assert.Nil(t, r)
-	})
-
-	t.Run("invalid memory request", func(t *testing.T) {
-		a := map[string]string{annotations.KeyCPURequest: "100m", annotations.KeyMemoryRequest: "memory"}
-		r, err := getResourceRequirements(a)
-		assert.NotNil(t, err)
-		assert.Nil(t, r)
-	})
-}
-
-func TestGetProbeHttpHandler(t *testing.T) {
-	pathElements := []string{"api", "v1", "healthz"}
-	expectedPath := "/api/v1/healthz"
-	expectedHandler := corev1.ProbeHandler{
-		HTTPGet: &corev1.HTTPGetAction{
-			Path: expectedPath,
-			Port: intstr.IntOrString{IntVal: SidecarHTTPPort},
-		},
-	}
-
-	assert.EqualValues(t, expectedHandler, getProbeHTTPHandler(SidecarHTTPPort, pathElements...))
-}
-
-func TestFormatProbePath(t *testing.T) {
-	testCases := []struct {
-		given    []string
-		expected string
-	}{
-		{
-			given:    []string{"api", "v1"},
-			expected: "/api/v1",
-		},
-		{
-			given:    []string{"//api", "v1"},
-			expected: "/api/v1",
-		},
-		{
-			given:    []string{"//api", "/v1/"},
-			expected: "/api/v1",
-		},
-		{
-			given:    []string{"//api", "/v1/", "healthz"},
-			expected: "/api/v1/healthz",
-		},
-		{
-			given:    []string{""},
-			expected: "/",
-		},
-	}
-
-	for _, tc := range testCases {
-		assert.Equal(t, tc.expected, formatProbePath(tc.given...))
-	}
-}
 
 func TestGetSidecarContainer(t *testing.T) {
 	t.Run("get sidecar container without debugging", func(t *testing.T) {
